requirements.txt
Flask
Flask-SocketIO
python-usernames
